Use NLTK stopwords as a provider
