+++ src/main/java/ch/uzh/rackrec/nlp/NLTKStopWordProvider.java
@@ -1,0 +1,146 @@
+package ch.uzh.rackrec.nlp;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+public class NLTKStopWordProvider implements IStopWordProvider {
+
+    public boolean isStopWord(String word) {
+        return getStopWords().contains(word);
+    }
+
+    public List<String> getStopWords() {
+
+        String[] stopWords = {
+            "i",
+            "me",
+            "my",
+            "myself",
+            "we",
+            "our", 
+            "ours", 
+            "ourselves", 
+            "you", 
+            "your", 
+            "yours", 
+            "yourself", 
+            "yourselves", 
+            "he", 
+            "him", 
+            "his", 
+            "himself", 
+            "she", 
+            "her", 
+            "hers", 
+            "herself", 
+            "it", 
+            "its", 
+            "itself", 
+            "they", 
+            "them", 
+            "their", 
+            "theirs", 
+            "themselves", 
+            "what", 
+            "which", 
+            "who", 
+            "whom", 
+            "this", 
+            "that", 
+            "these", 
+            "those", 
+            "am", 
+            "is", 
+            "are", 
+            "was", 
+            "were", 
+            "be", 
+            "been", 
+            "being", 
+            "have", 
+            "has", 
+            "had", 
+            "having", 
+            "do", 
+            "does", 
+            "did", 
+            "doing", 
+            "a",
+            "an", 
+            "the",
+            "and", 
+            "but", 
+            "if", 
+            "or", 
+            "because", 
+            "as", 
+            "until", 
+            "while", 
+            "of", 
+            "at", 
+            "by", 
+            "for", 
+            "with", 
+            "about", 
+            "against", 
+            "between", 
+            "into", 
+            "through", 
+            "during", 
+            "before", 
+            "after", 
+            "above", 
+            "below", 
+            "to", 
+            "from", 
+            "up", 
+            "down", 
+            "in", 
+            "out", 
+            "on", 
+            "off", 
+            "over", 
+            "under", 
+            "again", 
+            "further", 
+            "then", 
+            "once", 
+            "here", 
+            "there", 
+            "when", 
+            "where", 
+            "why", 
+            "how", 
+            "all", 
+            "any", 
+            "both", 
+            "each", 
+            "few", 
+            "more", 
+            "most", 
+            "other", 
+            "some", 
+            "such", 
+            "no", 
+            "nor", 
+            "not", 
+            "only", 
+            "own", 
+            "same",
+             "so", 
+            "than", 
+            "too", 
+            "very", 
+            "s",
+            "t",
+            "can",
+            "will",
+            "just",
+            "don",
+            "should"
+        };
+        List<String> stopWordList = new ArrayList<String>(Arrays.asList(stopWords));
+        return stopWordList;
+    }
+}
