+++ src/test/java/ch/uzh/rackrec/nlp/IdentifierLemmatizerTest.java
@@ -1,0 +1,4 @@
+package ch.uzh.rackrec.nlp;
+
+public class IdentifierLemmatizerTest {
+}
